+++ client/pages/auth/signin.vue
@@ -1,97 +1,21 @@
 <template>
   <vs-row>
-    <vs-col
-      vs-type="flex"
-      vs-justify="center"
-      vs-align="center"
-      vs-w="5.5"
-      vs-lg="5.5"
-      vs-sm="5.5"
-      vs-xs="0"
-    >
-      <div class="main-image">
-        <img src="~/assets/signin.png" />
-        <div class="contents">
-          <p class="title">Project</p>
-        </div>
-      </div>
-    </vs-col>
-    <vs-col
-      vs-type="flex"
-      vs-align="start"
-      vs-w="6.5"
-      vs-lg="6.5"
-      vs-sm="5.5"
-      vs-xs="12"
-    >
-      <vs-row>
-        <vs-col
-          vs-type="flex"
-          vs-justify="flex-end"
-          vs-lg="12"
-          vs-sm="12"
-          vs-xs="12"
-        >
-          <div class="nav">
-            <p>
-              既にアカウントをお持ちではない方は<nuxt-link to="/auth/signup"
-                >こちら</nuxt-link
-              >
-            </p>
-          </div>
-        </vs-col>
-        <vs-col vs-type="flex" vs-justify="center" vs-w="12">
-          <SigninForm />
-        </vs-col>
-      </vs-row>
-    </vs-col>
+    <GoToSignupLink />
+    <SigninForm />
   </vs-row>
 </template>
 
 <script lang="ts">
 import Vue from 'vue'
-import SigninForm from '@/components/SigninForm.vue'
+import SigninForm from '@/components/container/auth/signin/SigninForm.vue'
+import GoToSignupLink from '@/components/presentational/GoToSignupLink.vue'
 
 export default Vue.extend({
-  layout: 'guest',
+  layout: 'signup',
   middleware: 'guest',
   components: {
-    SigninForm
+    SigninForm,
+    GoToSignupLink
   }
 })
 </script>
-
-<style lang="scss" scoped>
-.nav {
-  p {
-    padding-top: 2vh;
-    padding-right: 2vw;
-    font-weight: normal;
-    font-family: HiraginoSans-W3;
-    color: #707070;
-    font-size: calc(8px + 0.5vw);
-  }
-}
-
-.main-image {
-  position: relative;
-  img {
-    width: 100%;
-    height: 100vh;
-    object-fit: cover;
-  }
-  .contents {
-    position: absolute;
-    font-weight: normal;
-    font-family: HiraginoSans-W3;
-    text-shadow: 0 3px 6px rgba(0, 0, 0, 0.2);
-    color: #707070;
-    top: 89px;
-    left: 94px;
-    .title {
-      font-size: 36px;
-      margin-bottom: 100px;
-    }
-  }
-}
-</style>
